+++ src/main/java/racing/domain/RandomGenerateNumber.java
@@ -1,4 +1,4 @@
-package racing;
+package racing.domain;
 
 import java.util.Random;
